Refactorized security controllers
