+++ src/Controller/Security/GoogleConnectStartController.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace App\Controller;
+namespace App\Controller\Security;
 
 use KnpU\OAuth2ClientBundle\Client\ClientRegistry;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -20,32 +20,24 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Security\Http\Util\TargetPathTrait;
 
-class SecurityGithubProviderController extends AbstractController
+/**
+ * @Route("/connect/google", name="connect_google_start")
+ */
+class GoogleConnectStartController extends AbstractController
 {
     use TargetPathTrait;
 
-    /**
-     * @Route("/connect/github", name="connect_github_start")
-     */
-    public function connect(Request $request, ClientRegistry $clientRegistry): Response
+    public function __invoke(Request $request, ClientRegistry $clientRegistry): Response
     {
         if ($targetPath = $request->query->get('_target_path')) {
             $this->saveTargetPath($request->getSession(), 'main', $targetPath);
         }
 
         return $clientRegistry
-            ->getClient('github')
+            ->getClient('google')
             ->redirect([
-                'user:email',
+                'openid', 'email', 'profile',
             ])
-            ;
-    }
-
-    /**
-     * @Route("/connect/github/check", name="connect_github_check")
-     */
-    public function check(): void
-    {
-        throw new \RuntimeException('This method should not be called.');
+        ;
     }
 }
